+++ docs/CHANGES.rst
@@ -14,6 +14,8 @@ Changelog
 
 Fixes:
 
+- Removed docstrings from PropertyManager methods to avoid publishing them.  [maurits]
+
 - Added publishing patch from Products.PloneHotfix20160419.
   This avoids publishing some methods inherited from Zope or CMF.  [maurits]